put UserRepository in separate package
